- feat: 장바구니 API/서비스 구축 및 CART 기반 주문 생성 연동

+++ src/main/java/com/jkomerce/store/service/OrderService.java
@@ -100,7 +100,7 @@ public class OrderService {
 
         List<OrderItemDTO> orderItems = new ArrayList<>();
 
-        int totalAmount = 0;
+        Long totalAmount = 0L;
         for(CartItemForOrderDTO item : items){
             Long itemId = item.getItemId();
             Integer quantity = item.getQuantity();
